Add priority and display name to context files

+++ app/controllers/context_categories_controller.rb
@@ -17,9 +17,11 @@ class ContextCategoriesController < ApplicationController
   def create
     file = params[:context_category][:context_file]
 
-    file_name = File.basename(file.original_filename ,File.extname(file.original_filename))
+    file_name = params[:context_category][:name]
+    return if file_name.empty?
+    priority = params[:context_category][:priority].present? ? params[:context_category][:priority] : 3
     ContextCategory.where(name: file_name).destroy_all
-    context = ContextCategory.create!(name: file_name)
+    context = ContextCategory.create!(name: file_name, priority: priority)
 
     file_contents = file.read
     file_contents.gsub!(/\r\n?/, "\n")
